requirements.txt
PySide6>=6.5.0
PyMuPDF>=1.23.0
paddleocr>=2.7.0
paddlepaddle>=2.5.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0
